// frontend-admin/src/api/generated/model/verticalProfileDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */

export interface VerticalProfileDto {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  name?: string | null;
  optionalFields?: unknown;
  posFeatures?: unknown;
  /** @nullable */
  posLayout?: string | null;
  requiredFields?: unknown;
}